requirements.txt
streamlit>=1.31.0
Pillow>=10.0.0
python-pptx>=0.6.21
google-generativeai>=0.3.0
openai>=1.12.0
gTTS>=2.5.0
elevenlabs>=0.3.0
streamlit-webrtc>=0.47.1
asyncio>=3.4.3
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
torchaudio>=2.0.0
soundfile>=0.12.1
phonemizer>=3.2.0
pdf2image>=1.16.3
PyPDF2>=3.0.0 